ENH: Fix implicit type conversion.

+++ src/iapws.h
@@ -16,7 +16,7 @@
  * @param[in] size_solvent Length of solvent string
  * @return kh Henry constant in mole fraction per GPa. NaN if gas not found.
  */
-extern double iapws_capi_kh(double T, char *gas, char *solvent, int size_gas, int size_solvent);
+extern double iapws_capi_kh(double T, char *gas, char *solvent, size_t size_gas, size_t size_solvent);
 
 /** 
  * @brief Compute the vapor-liquid constant for a given temperature and gas in solvent 
@@ -27,6 +27,6 @@ extern double iapws_capi_kh(double T, char *gas, char *solvent, int size_gas, in
  * @param[in] size_solvent Length of solvent string
  * @return kd Vapor-liquid constant. NaN if gas not found.
  */
-extern double iapws_capi_kd(double T, char *gas, char *solvent, int size_gas, int size_solvent);
+extern double iapws_capi_kd(double T, char *gas, char *solvent, size_t size_gas, size_t size_solvent);
 
 #endif
